Update class exclusion for dynamic compilation
