+++ osu.Game/Screens/ScorePresentType.cs
@@ -1,0 +1,11 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+namespace osu.Game.Screens
+{
+    public enum ScorePresentType
+    {
+        Results,
+        Gameplay
+    }
+}
